+++ ITechArtPizzaDelivery.Domain/Interfaces/IIngredientsService.cs
@@ -8,8 +8,8 @@ namespace ITechArtPizzaDelivery.Domain.Interfaces
     public interface IIngredientsService
     {
         Task<List<Ingredient>> GetAll();
-        Task<Ingredient> GetById(long id);
+        Task<Ingredient> GetById(int id);
         Task<Ingredient> Post(Ingredient ingredient);
-        Task DeleteById(long id);
+        Task DeleteById(int id);
     }
 }
