+++ ITechArtPizzaDelivery.Web/Models/PostPizzaModel.cs
@@ -10,7 +10,6 @@ namespace ITechArtPizzaDelivery.Web.Models
         public string Name { get; set; }
         public string Image { get; set; }
         public decimal Price { get; set; }
-        // change long to int
-        public long[] IngredientsId { get; set; }
+        public int[] IngredientsId { get; set; }
     }
 }
